% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_to_rel.R
\name{abs_to_rel}
\alias{abs_to_rel}
\title{Convert absolute path to relative one}
\usage{
abs_to_rel(path, base, mustWork = NA)
}
\arguments{
\item{path}{\itemize{
\item path to convert to relative to base
}}

\item{base}{\itemize{
\item base folder for relative path
}}

\item{mustWork}{\itemize{
\item logical: if TRUE then an error is given if the
result cannot be determined; if NA then a warning.
}}
}
\value{
relative path as string
}
\description{
Function to convert an absolute path to relative one.
}
\details{
See \link{normalizePath} for more information on \code{mustWork}. The code
for this function is essentially a copy of that for \link[sen2r:abs2rel]{sen2r::abs2rel}.
}
