% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextFunctions.R
\name{addQuotes}
\alias{addQuotes}
\title{Add quotes to a character string vector.}
\usage{
addQuotes(str, qt = "'")
}
\arguments{
\item{str}{\itemize{
\item character vector to add quotes to
}}

\item{qt}{\itemize{
\item quote character
}}
}
\value{
character vector w/ elements in quotes
}
\description{
Function to add quotes to a character string vector.
}
\details{
None.
}
