% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateDataframe.R
\name{aggregateDataframe}
\alias{aggregateDataframe}
\title{Aggregate a dataframe according to "factor" levels.}
\usage{
aggregateDataframe(dfr, factors = NULL, vars = NULL, verbose = FALSE)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe
}}

\item{factors}{\itemize{
\item factors to aggregate by
}}

\item{vars}{\itemize{
\item variable(s) to aggregate
}}
}
\value{
dataframe with vars aggregated by levels of "factor" columns
}
\description{
Function to aggregate a dataframe according to "factor" levels.
}
\details{
Uses sqldf.
}
