% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxify.R
\name{boxify}
\alias{boxify}
\title{"Boxify" x,y vectors for a "boxy" xy plot}
\usage{
boxify(x, y, dx = NULL)
}
\arguments{
\item{x}{-}

\item{y}{-}

\item{dx}{}
}
\value{
dataframe with expanded elements x,y
}
\description{
Function to "boxify" x,y vectors for a "boxy" xy plot.
}
\details{
This function expands each x,y pair to x+c(-1,0,0,1)*dx/2 and
y*c(0,1,1,0), which provides a "box".
}
