% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCIs.R
\name{calcCIs}
\alias{calcCIs}
\title{Compute confidence intervals from cv's or standard deviations}
\usage{
calcCIs(vals, cvs = NULL, sdvs = NULL, pdfType = c("lognormal", "normal"),
  ci = 0.95, verbose = FALSE)
}
\arguments{
\item{vals}{- vector of values}

\item{cvs}{- vector of cvs}

\item{sdvs}{- vector of arithmetic-scale standard deviations}

\item{pdfType}{- probability distribution for error bars}

\item{ci}{- confidence interval for error bar plots}

\item{verbose}{- flag (T/F) to print intermediate output}
}
\value{
list with vectors lci, uci as  named elements
}
\description{
Function to compute confidence intervals.
}
\details{
None.
}
