% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLnScaleStdDevs.R
\name{calcLnScaleStdDevs}
\alias{calcLnScaleStdDevs}
\title{Compute ln-scale standard deviations from cv's}
\usage{
calcLnScaleStdDevs(cvs)
}
\arguments{
\item{cvs}{\itemize{
\item vector of cv's
}}
}
\value{
vector of ln-scale standard deviations
}
\description{
Function to compute ln-scale standard deviations from cv's.
}
\details{
Conversion from \code{cv} to ln-scale standard deviation \code{lnSD} is

lnSD = sqrt(log(1+cv^2))
}
