% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMidpoints.R
\name{calcMidpoints}
\alias{calcMidpoints}
\title{Calculate midpoints from a vector ofcuutpoints}
\usage{
calcMidpoints(x)
}
\arguments{
\item{x}{\itemize{
\item numeric vector of cutpoints
}}
}
\value{
vector of midpoints (length = length(x)-1)
}
\description{
Funciton to calculate midpoints from a vector ofcuutpoints
}
\details{
calculates the arithmetic midpoints of the successive intervals
defined by x.
}
