% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidType.R
\name{checkValidType}
\alias{checkValidType}
\title{Check a value (type) against a vector of valid types}
\usage{
checkValidType(type, valid_types)
}
\arguments{
\item{type}{\itemize{
\item string to check
}}

\item{valid_types}{\itemize{
\item vector of valid types
}}
}
\value{
TRUE if the value of \code{type} is a valid type, FALSE otherwise.
}
\description{
Function to check type against vector of valid types.
}
\details{
If FALSE, a warning message is generated.
It is the responsibility of the caller to determine what an
appropriate response is.
}
