% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectValuesByGroup.R
\name{collectValuesByGroup}
\alias{collectValuesByGroup}
\title{Collect values in a dataframe by implicit group}
\usage{
collectValuesByGroup(
  dfr0,
  collect = "y",
  names_from = "z",
  values_from = "val"
)
}
\arguments{
\item{dfr0}{\itemize{
\item dataframe on which to operate
}}

\item{collect}{\itemize{
\item name(s) of columns from which to collect values
}}

\item{names_from}{\itemize{
\item column name with names to pivot wider (optional)
}}

\item{values_from}{\itemize{
\item column name with values to pivot wider (optional)
}}
}
\description{
Function to collect values from a column in a dataframe into a list vector by group.
}
\details{
This function optionally pivots an input dataframe wider using \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}
with \code{names_from} and \code{values_from}. It then selects only the distinct rows in the resulting dataframe
after the column(s) identified by \code{collect} are dropped, and assigns a \code{group} number to each distinct row.
The values in the \code{collect} column(s) of the input dataframe are then identified by group and "collected"
by group membership into a vector. The list of "collected" vectors is then appended to the "distinct" dataframe.
}
