% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeStats.dataframe.R
\name{computeStats.dataframe}
\alias{computeStats.dataframe}
\title{computeStats.dataframe: Calculate count, mean, stdev, and cv of selected columns of a dataframe}
\usage{
computeStats.dataframe(x, cols.factors = 1, cols.vars = 2)
}
\arguments{
\item{x}{- a dataframe}

\item{cols.factors}{- vector identifying columns to use as factors}

\item{cols.vars}{- vector identifying columns to use as variables}
}
\value{
- NULL (not yet implemented)
}
\description{
Returns count, mean, stdev, and cv of a dataframe
}

