% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStats.matrix.R
\name{computeStats.matrix}
\alias{computeStats.matrix}
\title{computeStats.matrix: Calculate count, mean, stdev, and cv of the rows/columns of a matrix}
\usage{
computeStats.matrix(x, byCol = TRUE)
}
\arguments{
\item{x}{\itemize{
\item matrix
}}

\item{byCol}{\itemize{
\item compute stats by column (if TRUE)
}}
}
\value{
\itemize{
\item for x a matrix, returns matrix w/ rows with elements count, mean, stdev and cv.
Row names correspond to column (or row) names of original matrix.
}
}
\description{
Returns count, mean, stdev, and cv of the rows or columns of a matrix
}
