% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTickInterval.R
\name{computeTickInterval}
\alias{computeTickInterval}
\title{Calculate nice tick intervals for a scale.}
\usage{
computeTickInterval(rng, nticks = 5)
}
\arguments{
\item{rng}{\itemize{
\item the nominal range for the scale \link{i.e., c(min,max)}
}}

\item{nticks}{\itemize{
\item the desired number of ticks
}}
}
\description{
Function to calculate nice tick intervals for a scale
based on a range and the desired number of ticks.
}
\details{
None.
}
