% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateCSVs.R
\name{concatenateCSVs}
\alias{concatenateCSVs}
\title{Concatenate csv files into one}
\usage{
concatenateCSVs(
  caption = "Select csv files",
  out.csv = "test.csv",
  out.dir = NULL,
  method = "writeLines"
)
}
\arguments{
\item{caption}{\itemize{
\item caption for file dialog box
}}

\item{out.csv}{\itemize{
\item filename of output csv file
}}

\item{out.dir}{\itemize{
\item directory for output csv file
}}

\item{method}{\itemize{
\item specifies method used for concatenation ('writeLines' or 'write.table')
}}
}
\description{
Function to concatenate a number of csv files (with identical column structure)
into a single csv file.
}
\details{
None.
}
