% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToNumeric.R
\name{convertToNumeric}
\alias{convertToNumeric}
\title{Convert a character vector to numeric values}
\usage{
convertToNumeric(x)
}
\arguments{
\item{x}{\itemize{
\item vector of values on which to perform conversion
}}
}
\value{
a vector of the same length as \code{x} converted to numeric
}
\description{
Function to convert a character vector to numeric values.
}
\details{
Values with commas are converted after stripping out commas.
Values with percent signs are converted after removing percent signs and
converting to fractions (i.e., dividing by 100). Otherwise, this works as \code{as.numeric(v)} would.
}
