% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyFiles.R
\name{copyFiles}
\alias{copyFiles}
\title{Copy new or updated files/folders}
\usage{
copyFiles(dirFrom = NULL, dirTo = NULL)
}
\arguments{
\item{dirFrom}{\itemize{
\item path to folder to copy files/folders from
}}

\item{dirTo}{\itemize{
\item path to folder to copy files/folders to
}}
}
\value{
tibble of new or updated files/folder paths in the "to" folder
}
\description{
Function to copy new or updated files and folders from one
location to another.
}
\details{
File/folder lists are recursively compiled into tibbles for both "from" and "to"
folders and the resulting "from" tibble is left-joined to the "to" tibble based on the
folder/file names under the base folders. The file modification times are checked for
files that appear in both file structures and new files or files with with later modification times in the
"from" folders are copied into their respective "to" folders. Folders in "from" but missing
in "to" are created in "to" and any files are copied into the new folders in "to".

If \code{dirFrom} or \code{dirTo} is NULL (the default for both), a browser is opened allowing the user to select
the folder by selecting \strong{one of the files} it contains.
}
