% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Returns count of x (ignoring NAs)}
\usage{
count(x)
}
\arguments{
\item{x}{\itemize{
\item a vector of numbers (or object coerceable to a vector)
}}
}
\value{
number of non-NAs in x
}
\description{
Returns count of x (ignoring NAs)
}
