% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormulas.R
\name{createFormulas}
\alias{createFormulas}
\title{Create a set of formulas from combinations of terms in one}
\usage{
createFormulas(frmla, keep = NULL)
}
\arguments{
\item{frmla}{\itemize{
\item base formula to expand
}}

\item{keep}{\itemize{
\item NULL, or character vector of terms which must be included in final set
}}
}
\value{
list of formulas
}
\description{
Function to create a set formulas from combinations of terms.
}
\details{
If \code{keep} is NULL (the default), then the list of returned formulas includes all
possible combinations of terms. If \code{keep} is not NULL, then the returned list
only includes terms involving those specified in \code{keep}.
}
