% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGridForCategories.R
\name{createGridForCategories}
\alias{createGridForCategories}
\title{Create a \pkg{tibble} tibble describing a rectangular grid defined by two sets of categories}
\usage{
createGridForCategories(xcats, ycats)
}
\arguments{
\item{xcats}{\itemize{
\item vector of categories for x-axis
}}

\item{ycats}{\itemize{
\item vector of categories for y-axis
}}
}
\value{
a \code{\link[tibble]{tibble}}. See Details.
}
\description{
Function to create a \pkg{tibble} tibble describing a rectangular grid defined by two sets of categories.
}
\details{
The returned tibble will have the following columns:
\itemize{
\item{xc - x-axis category labels (factor levels)}
\item{yc - y-axis category labels (factor levels)}
\item{id - id (character) for rectangle with center xc,yc [="(xc,yc)"]}
\item{x - rectangle corner location (numeric) along x-axis}
\item{y - rectangle corner location (numeric) along y-axis}
}
}
