% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvp.R
\name{cvp}
\alias{cvp}
\title{Calculate the cv (or std dev, if mean=0) of a vector.}
\usage{
cvp(x)
}
\arguments{
\item{x}{\itemize{
\item vector
}}
}
\value{
cv (or std dev, if mean=0).
}
\description{
Function to calculate the cv (or std dev, if mean=0) of a vector.
}
\details{
Uses package \code{stats}.
}
