% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCol.R
\name{deleteCol}
\alias{deleteCol}
\title{Delete column in a dataframe or matrix}
\usage{
deleteCol(tbl, col)
}
\arguments{
\item{tbl}{- dataframe or matrix from which to delete column}

\item{col}{- name of column to delete}
}
\value{
dataframe or matrix with column removed
}
\description{
Function to delete column in a dataframe or matrix
}
