% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCols.R
\name{deleteCols}
\alias{deleteCols}
\title{Delete columns in a dataframe or matrix}
\usage{
deleteCols(tbl, cols, debug = FALSE)
}
\arguments{
\item{tbl}{\itemize{
\item dataframe or matrix from which to delete columns
}}

\item{cols}{\itemize{
\item names of column to delete
}}

\item{debug}{\itemize{
\item T/F to print debugging info
}}
}
\value{
dataframe or matrix with columns removed
}
\description{
Function to delete columns in a dataframe or matrix
}
