% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropLevels.R
\name{dropLevels}
\alias{dropLevels}
\title{Drop factor levels from a dataframe}
\usage{
dropLevels(tbl, dropLevels = NULL)
}
\arguments{
\item{tbl}{- dataframe to drop levels from}

\item{dropLevels}{- list by factor name of factor levels to drop}
}
\value{
dataframe
}
\description{
Function to drop factor levels from a dataframe.
}
\details{
Distinct levels of each factor can be dropped from the
exported dataframe by seting dropLevels to a list with names corresponding to
factor columns and values being vectors of factor levels to drop.
}
