% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatZeros.R
\name{formatZeros}
\alias{formatZeros}
\title{formatZeros: replace blanks with 0's in printed numbers}
\usage{
formatZeros(x, width = 2, format = "d", ...)
}
\arguments{
\item{x}{- object that can be coerced to a numeric vector using as.numeric(as.vector(x))}

\item{width}{- width of formatted numbers}

\item{format}{- c-type format string (see "format" description in base R)}

\item{...}{- additional inputs to "format"}
}
\value{
character vector with blanks replaced by zeros
}
\description{
Format input vector, replacing leading or trailing blanks with zeros.
}
\details{

}

