% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomMean_HarmMean.R
\name{GeomMean}
\alias{GeomMean}
\title{Calculate geometric mean with NAs removed}
\usage{
GeomMean(x, verbose = FALSE)
}
\arguments{
\item{x}{\itemize{
\item vector
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a numeric value.
}
\description{
Function to calculate geometric mean with NAs removed.
}
\details{
If x has any 0's, then 0 is returned. Otherwise exp((mean(log(x),na.rm=TRUE))).
}
