% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCaptionsFromGGPlotList.R
\name{getCaptionsFromGGPlotList}
\alias{getCaptionsFromGGPlotList}
\title{Extract captions from list of plots in printGGList format}
\usage{
getCaptionsFromGGPlotList(plts)
}
\arguments{
\item{x}{- list of plots in printGGList format, with captions as list names}
}
\value{
character vector with plot captions as elements
}
\description{
Function to extract captions from list of plots in printGGList format.
}
\details{
Uses \code{stringr::str_trim} to trim leading and trailing whitespace.
}
