% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFactorLevels.R
\name{getFactorLevels}
\alias{getFactorLevels}
\title{Find the unique combination of levels of a set of "factors" in a dataframe}
\usage{
getFactorLevels(
  dfr,
  factors,
  sortBy = NULL,
  ascending = TRUE,
  orthogonal = FALSE,
  debug = FALSE
)
}
\arguments{
\item{dfr}{\itemize{
\item input dataframe (or object that can be cast to a dataframe)
}}

\item{factors}{\itemize{
\item vector of desired factor names from dfr
}}

\item{sortBy}{\itemize{
\item vector of factor names to sort by (in order of precedence)
}}

\item{ascending}{\itemize{
\item flag (T/F) to sort in ascending/descending order
}}

\item{orthogonal}{\itemize{
\item flag (T/F) to form orthogonal combination of individual factors
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
a dataframe containing each unique combination of factor levels in dfr
}
\description{
Returns the unique combination of levels
of the specified "factors" in a dataframe.
}
