% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileExtension.R
\name{getFileExtension}
\alias{getFileExtension}
\title{Get the extension from a filename}
\usage{
getFileExtension(fn)
}
\arguments{
\item{fn}{\itemize{
\item the filename
}}
}
\value{
\itemize{
\item the extension (without the "."), or NA if no extension
}
}
\description{
Function to get the extension from a filename
}
\details{
If the filename has no extension, the functions returns NA. If the filename ends
with ".", then "" is returned.
}
