% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFile-FolderStructure.R
\name{getFileStructure}
\alias{getFileStructure}
\title{Get file structure}
\usage{
getFileStructure(
  path = ".",
  all.files = TRUE,
  full.names = TRUE,
  recursive = TRUE,
  include.dirs = FALSE
)
}
\arguments{
\item{path}{- path to toplevel folder}

\item{all.files}{- flag to include all files}

\item{full.names}{- flag to return fullnames}

\item{recursive}{- flag to determine structure recursively}

\item{include.dirs}{- flag to include folders}
}
\value{
text vector
}
\description{
Get file structure
}
\details{
Uses \code{[base]{list.files}} to recursively determine file structure.
}
