% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFile-FolderStructure.R
\name{getFolderStructure}
\alias{getFolderStructure}
\title{Get folder structure}
\usage{
getFolderStructure(path = ".", full.names = TRUE, recursive = TRUE)
}
\arguments{
\item{path}{- path to toplevel folder}

\item{full.names}{- flag to return fullnames}

\item{recursive}{- flag to determine structure recursively}
}
\value{
text vector
}
\description{
Get folder structure
}
\details{
Uses \code{[base]{list.dirs}} to determine folder structure.
}
