% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntervalLimits.R
\name{getIntervalLimits}
\alias{getIntervalLimits}
\title{Get interval limits (cut points) from a vector of midpoints}
\usage{
getIntervalLimits(xs)
}
\arguments{
\item{xs}{- the vector of interval midpoints}
}
\value{
the vector of interval limits (cutpoints)
}
\description{
Function to return a vector of interval limits from a vector of midpoints.
}
\details{
Assumes interval limits are midway between coordinates in xs. First
and last coordinates assume same widths as first and last cells. Resulting vector
has nx+1 values, if xs has nx values.
}

