% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObj_saveObj.R
\name{getObjNames}
\alias{getObjNames}
\title{Get the names of objects saved in an RData file}
\usage{
getObjNames(fn)
}
\arguments{
\item{fn}{\itemize{
\item filename from which to extract the names
}}
}
\value{
a character vector with the object names
}
\description{
Function to get the names of objects saved in an RData file.
}
\details{
The RData file is "loaded" in the function environment to determine
the names of the associated objects.
}
