% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOperatingSystem.R
\name{getOperatingSystem}
\alias{getOperatingSystem}
\title{Get the operating system name.}
\usage{
getOperatingSystem()
}
\value{
Either 'OSX' or 'Windows'
}
\description{
Function that returns the operating system name.
}
\details{
Checks \code{Sys.info()[['sysname']]} for platform name and returns 'Win'
if lower case platform name is 'windows', 'OSX' if 'darwin'.
}
