% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_CreateColors.R
\name{gg_CreateColors}
\alias{gg_CreateColors}
\title{Create a vector of colors based on a color scale}
\usage{
gg_CreateColors(n, scale = "hue", ..., show = FALSE)
}
\arguments{
\item{n}{\itemize{
\item number of colours to create
}}

\item{scale}{\itemize{
\item name of scale to use (see Details)
}}

\item{...}{\itemize{
\item passed on to the scale
}}

\item{show}{\itemize{
\item show colours using \code{\link[scales:show_col]{scales::show_col()}}
}}
}
\value{
a vector of colors.
}
\description{
Function to create a vector of colors based on a color scale.
}
\details{
Available color scales (and default parameters) are\cr
\itemize{
\item{"hue" - ...: none (see \code{\link[scales:hue_pal]{scales::hue_pal()}})}
\item{"brewer" - ...: type="seq", palette=1, direction=1 (see \code{\link[scales:brewer_pal]{scales::brewer_pal()}})}
\item{"viridis" - ...: alpha=1,begin=0,end=0,direction=1,option='D' (see \code{\link[scales:viridis_pal]{scales::viridis_pal()}})}
}
}
\examples{
gg_CreateColors(3,show=TRUE);
gg_CreateColors(3,"brewer",show=TRUE);
gg_CreateColors(3,"viridis",show=TRUE); 

}
\seealso{
\code{\link[scales:hue_pal]{scales::hue_pal()}}, \code{\link[scales:brewer_pal]{scales::brewer_pal()}},\code{\link[scales:viridis_pal]{scales::viridis_pal()}}
}
