% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GetPlotDims.R
\name{gg_GetPlotDims}
\alias{gg_GetPlotDims}
\title{Get dimensions of a ggplot2 object or derived gtable}
\usage{
gg_GetPlotDims(
  p,
  unit = "in",
  null = 3,
  null_x = NULL,
  null_y = NULL,
  returnAll = FALSE,
  verbose = 0
)
}
\arguments{
\item{p}{\itemize{
\item ggplot2 or derived gtable object
}}

\item{unit}{\itemize{
\item plot units ("in", "cm", etc.)
}}

\item{null}{\itemize{
\item size of (x and y) null (resizeable) units (in plot units; overridden if null_x and null_y are given)
}}

\item{null_x}{\itemize{
\item size of horizontal null (resizeable) units (in plot units; must also specify null_y)
}}

\item{null_y}{\itemize{
\item size of vertical (resizeable) units (in plot units; must also specify null_x)
}}

\item{returnAll}{\itemize{
\item flag to return all information (not just dimensions)
}}

\item{verbose}{\itemize{
\item flag (0,1,2) to print intermediate information (0=off)
}}
}
\value{
if returnAll is FALSE, returns a vector with plot width and height. Otherwise, it
invisibly returns a list with elements
dims - vector with plot width and height
lyt - layout table
tbl_rows - row dimensions
tbl_cols - column dimensions
tbl_matp -tibble with intermediate information
tbl_matpp - tibble with intermediate information
}
\description{
Function to get dimensions of a ggplot2 object or derived gtable.
}
\details{
This function converts "null" (relative) scales in a ggplot2 object
(or derived gtable) to absolute scales.

If either null_x or null_y is NULL, then the value of \code{null} is used for both. If
both null_x and null_y are not NULL, these values are used for nulls along the corresponding
axes.
}
\note{
This does not work (yet) for composite figures consisting of multiple ggplot2 objects.
}
\author{
William Stockhausen (william.stockhausen@noaa.gov)
}
