% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GetPlotLayout.R
\name{gg_GetPlotLayout}
\alias{gg_GetPlotLayout}
\title{Get the layout dataframe from a ggplot2 or gtable object}
\usage{
gg_GetPlotLayout(p, keepAll = TRUE, verbose = 0)
}
\arguments{
\item{p}{- ggplot2 or gtable object}

\item{keepAll}{- flag (T/F) to keep all grobs (including grobs with physical size 0)}

\item{verbose}{- flag (T/F) to print intermediate information}
}
\value{
returns a dataframe with columns reflecting the row/column locations and extents
of the input ggplot2 or gtable object.
}
\description{
Function to get the layout dataframe from a ggplot2 or gtable object.
}
\details{
This function adds information to a layout regarding how grobs
are assigned to rows and columns in a ggplot2 (gtable) layout. It adds coordinates to
allow plotting grobs as rectangles by row/column coordinates.
}
\author{
William Stockhausen (william.stockhausen@noaa.gov)
}
