% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_SavePlotWithOneFixedDim.R
\name{gg_SavePlotWithOneFixedDim}
\alias{gg_SavePlotWithOneFixedDim}
\title{Save a ggplot2 plot to a file, specifying the size of one dimension}
\usage{
gg_SavePlotWithOneFixedDim(
  filename,
  plot = last_plot(),
  fxdDim,
  fitWidth = TRUE,
  units = "in",
  device = NULL,
  dpi = 200
)
}
\arguments{
\item{filename}{- file name for output}

\item{plot}{- ggplot2 plot object to save (defaults to last plot)}

\item{fxdDim}{- value for the fixed dimension}

\item{fitWidth}{- flag indicating fixed dimension is width (default is TRUE)}

\item{units}{- character string with units for the fixed dimension}

\item{device}{- graphics device to print to}

\item{dpi}{- resolution (default=200 dpi)}
}
\description{
Function to save a ggplot2 plot to a file, 
specifying the size of one dimension.
}
\details{
Uses [gg_FixOnePlotDim] to determine the value of the missing plot dimension.
}
