% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ShowPlotGrobs.R
\name{gg_ShowPlotGrobs}
\alias{gg_ShowPlotGrobs}
\title{Show component "grobs" of a ggplot2 object (and get info on grob sizes)}
\usage{
gg_ShowPlotGrobs(
  p,
  unit = "in",
  xOffset = -3,
  yOffset = -3,
  showAll = FALSE,
  showPlot = TRUE,
  verbose = 0
)
}
\arguments{
\item{p}{- ggplot or gtable object}

\item{unit}{- plot units ("in", "cm", etc.)}

\item{xOffset}{- x-axis offset for row size labels}

\item{yOffset}{- y-axis offset for column size labels}

\item{showAll}{- flag (T/F) to show all grobs (including grobs with physical size 0)}

\item{showPlot}{- flag (T/F) to print grob plot}

\item{verbose}{- flag (0,1,2) to print intermediate information (0=off)}
}
\value{
invisibly returns a list with elements
plot - plot of object components with axes reflecting inferred sizes
lyt - vector with plot width and height
tbl_rows - 
sum_rows - 
tbl_cols - 
sum_cols -
}
\description{
Function to show component "grobs" of a ggplot2 object (and get info on grob sizes).
}
\details{
This function identifies scales of components in a ggplot2 object 
and creates an unfilled rectangle plot of the resulting components with axes reflecting
the rows and columns in the gtable derived from the ggplot2 object.

See also [gg_ShowPlotDims], which is similar but provides absolute scales.
}
\author{
William Stockhausen (william.stockhausen@noaa.gov)
}
