% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomMean_HarmMean.R
\name{HarmMean}
\alias{HarmMean}
\title{Calculate harmonic mean with NAs removed}
\usage{
HarmMean(x, verbose = FALSE)
}
\arguments{
\item{x}{\itemize{
\item vector
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a numeric value.
}
\description{
Function to calculate harmonic mean with NAs removed.
}
\details{
If x has any 0's, then 0 is returned. Otherwise 1/(mean(1/x,na.rm=TRUE)).
}
