% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBlankString.R
\name{isBlankString}
\alias{isBlankString}
\title{Test if character string is only whitespace (i.e., blank)}
\usage{
isBlankString(str)
}
\arguments{
\item{strv}{- character string to test}
}
\value{
TRUE or FALSE.
}
\description{
Function to test if character string is only whitespace (i.e., blank).
}
\details{
Returns TRUE if \code{stringr::str_squish(str)} reduces \code{str} to "", otherwise FALSE. 
Also returns FALSE if \code{str} is not a length 1 character vector.
}
