% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOperatingSystem.R
\name{isOSX}
\alias{isOSX}
\title{Test if operating system is OSX.}
\usage{
isOSX()
}
\value{
TRUE if the operating system is OSX, FALSE otherwise
}
\description{
Function that returns TRUE if the operating system is OSX, FALSE otherwise.
}
\details{
Checks \code{Sys.info()[['sysname']]} for platform name and returns TRUE
if the lower case platform name is 'darwin', FALSE otherwise.
}
