% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidDate.R
\name{isValidDate}
\alias{isValidDate}
\title{Test if object is a valid date}
\usage{
isValidDate(dt)
}
\arguments{
\item{dt}{\itemize{
\item the object to test
}}
}
\value{
logical
}
\description{
Function to test if an object is a valid date.
}
\details{
if is.na(as.Date(dt)) does not throw an error and is TRUE,
then the function returns TRUE; otherwise, it returns FALSE.

Based on discussion at https://gist.github.com/micstr/69a64fbd0f5635094a53.
}
\examples{
isValidDate("2023-02-28");  #--expect TRUE

isValidDate("2023-02-30"); #--expect FALSE

}
