% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOperatingSystem.R
\name{isWin}
\alias{isWin}
\title{Test if operating system is Windows.}
\usage{
isWin()
}
\value{
TRUE if the operating system is Windows, FALSE otherwise
}
\description{
Function that returns TRUE if the operating system is Windows, FALSE otherwise.
}
\details{
Checks \code{Sys.info()[['sysname']]} for platform name and returns TRUE
if the lower case platform name is 'windows', FALSE otherwise.
}
