% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepCols.R
\name{keepCols}
\alias{keepCols}
\title{Keep columns from a dataframe}
\usage{
keepCols(tbl, cols, debug = FALSE)
}
\arguments{
\item{tbl}{\itemize{
\item dataframe with columns to keep or discard
}}

\item{cols}{\itemize{
\item vector of column names to keep
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
a dataframe with columns to be kept, or NULL if no column names matched.
}
\description{
Function to keep columns from a dataframe
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  Creates a dataframe containing the specified subset of columns of input dataframe "tbl" by
  keeping only columns matching the specified names.
}\if{html}{\out{</div>}}
}
