% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPackageFiles.R
\name{listPackageFiles}
\alias{listPackageFiles}
\title{Get a vector of filenames associated with a package}
\usage{
listPackageFiles(pkg, sub = NULL)
}
\arguments{
\item{pkg}{\itemize{
\item name of package
}}

\item{sub}{\itemize{
\item (optional) name of subfolder to check
}}
}
\value{
a character vector of file paths relative to the package folder or sub-folder.
}
\description{
Function to get a vector of filenames associated with a package.
}
\details{
If \code{sub} is NULL, the names of all folders/files included in the package are returned.
Note, though, that this does NOT return file names from the package source, but from the compiled
package.
}
