% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.lapply.R
\name{my.lapply}
\alias{my.lapply}
\title{my version of lapply}
\usage{
my.lapply(X, FUN, ...)
}
\arguments{
\item{X}{\itemize{
\item data to apply function to (coerced to list if necessary)
}}

\item{FUN}{\itemize{
\item function to apply to X
}}

\item{...}{\itemize{
\item other inputs to FUN
}}
}
\value{
result of lapply'ing function to X
}
\description{
A convenience implementation of lapply
}
\details{
None.
}
