\name{parseDays}
\alias{parseDays}
\title{Convert character vector of dates to numeric vector of days-of-month.}
\usage{
parseDays(x, format = "MM/DD/YYYY")
}
\arguments{
  \item{x}{- the character vector to convert}

  \item{format}{- the date format (e.g.
  'MM/DD/YYYY','DD-MON-YYYY')}
}
\value{
numeric vector of day of month (1-31)
}
\description{

}

