% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause}
\alias{pause}
\title{Pause R for a defined time}
\usage{
pause(t, debug = FALSE)
}
\arguments{
\item{t}{- number of seconds to pause}

\item{debug}{- flag (T/F) to print debugging info}
}
\description{
Runs a loop that causes the system to pause operation for 't' seconds
}

