% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{reclassify}
\alias{reclassify}
\title{Reclassify values in a vector.}
\usage{
reclassify(x, rcl)
}
\arguments{
\item{x}{- vector with values to reclassify}

\item{rcl}{- list of lists specifying reclassifications}
}
\value{
vector of length(x)
}
\description{
Function to reclassify values in a vector
}
\details{
"rcl" is a list of lists. Each sublist is of the form list(new=new, old=old),
where the value "new" is the new classification assigned to each element in "x" 
that is found in the vector "old".
}

