% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify.R
\name{rectify}
\alias{rectify}
\title{"Rectify" x,y vectors for a ggplot2::geom_rect xy plot}
\usage{
rectify(x, y, dx = NULL)
}
\arguments{
\item{x}{\itemize{
\item vector to expand
}}

\item{y}{\itemize{
\item vector to expand
}}

\item{dx}{\itemize{
\item increment along x-axis
}}
}
\value{
dataframe with expanded elements xmin,xmax,ymin,ymax
}
\description{
Function to "rectxify" x,y vectors for a ggplot2::geom_rect xy plot.
}
\details{
None
}
