% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rootMedianSquare.R
\name{rootMedianSquare}
\alias{rootMedianSquare}
\title{Compute the root median square of a vector}
\usage{
rootMedianSquare(y)
}
\arguments{
\item{y}{- vector to compute rmds for}
}
\value{
the rmds
}
\description{
Function to compute the root median square of a vector
}
\details{

}

