% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObj_saveObj.R
\name{saveObj}
\alias{saveObj}
\title{Save an R object to an RData file with name 'obj'}
\usage{
saveObj(obj, fn)
}
\arguments{
\item{obj}{\itemize{
\item object to save
}}

\item{fn}{\itemize{
\item filename to which to save the R object
}}
}
\description{
Function to save an R object to an RData file with name 'obj'.
}
\details{
The original R object is saved with the name 'obj', so it can easily
be returned using the function \code{\link{getObj}}.
}
