% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFile.R
\name{selectFile}
\alias{selectFile}
\title{Select a file}
\usage{
selectFile(
  ext = "*",
  caption = paste0("Select .", ext, " file to import", ),
  path = rstudioapi::getActiveProject()
)
}
\arguments{
\item{ext}{\itemize{
\item extension for files to choose from
}}

\item{caption}{\itemize{
\item caption for file dialog (if file name not provided)
}}

\item{path}{\itemize{
\item path to folder (defaults to \code{\link[rstudioapi:getActiveProject]{rstudioapi::getActiveProject()}})
}}
}
\value{
Selected file name. Returns NULL if the user canceled selection using the file dialog.
}
\description{
Function allows the user to select a file using a gui interface.
}
\details{
Uses \code{\link[rstudioapi:file-dialogs]{rstudioapi::selectFile()}} to select file.
}
