% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortBy.R
\name{sortBy}
\alias{sortBy}
\title{sort a dataframe by row according to columns.}
\usage{
sortBy(df, factors, na.last = TRUE, ascending = TRUE, debug = FALSE)
}
\arguments{
\item{df}{\itemize{
\item input dataframe (or object that can be cast to a dataframe)
}}

\item{factors}{\itemize{
\item vector of desired column names from df to sort by
}}

\item{na.last}{\itemize{
\item flag (T/F) to order NAs last/first (=NA removes any NAs)
}}

\item{ascending}{\itemize{
\item flag (T/F) to sort in ascending/descending order
}}

\item{debug}{\itemize{
\item flag (T/F) to print debugging info
}}
}
\value{
sorted version of df
}
\description{
Returns a dataframe sorted by row according to the combination of values in
the columns specified by "factors".
}
\details{
None.
}
