% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.vector}
\alias{standardize.vector}
\title{Standardize a vector by converting to z-scores.}
\usage{
standardize.vector(x)
}
\arguments{
\item{x}{- vector to standardize}
}
\value{
standardized vector as z-scores (same size as x)

not exported
}
\description{
Function to standardize a vectorby converting to z-scores.
}
\details{

}

