% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum.R
\name{Sum}
\alias{Sum}
\title{Sum with NAs removed}
\usage{
Sum(x)
}
\arguments{
\item{x}{- object to be summed}
}
\description{
Convenience function to sum with NAs removed.
}
\details{
\code{Sum(x)} is equivalent to \code{sum(x,na.rm=TRUE)}
}
