% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumByDims.R
\name{sumByDims}
\alias{sumByDims}
\title{Sum array by specified dimensions}
\usage{
sumByDims(n, dims = 1)
}
\arguments{
\item{n}{- array to sum}

\item{dims}{- vector of indices for dimension to sum by}
}
\value{
array with dimensions used for summing
}
\description{
Function to sum array by specified dimensions.
}
\details{
Uses the \code{apply} function to sum over the requested dimension.
}

