% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimByResiduals.R
\name{trimByResiduals}
\alias{trimByResiduals}
\title{Trim rows from a dataframe based on a vector of residuals.}
\usage{
trimByResiduals(tbl, resids, n = 5, debug = TRUE)
}
\arguments{
\item{tbl}{\itemize{
\item dataframe with original data. row names should be 1:nrows BEFORE residuals are calculated
}}

\item{resids}{\itemize{
\item vector of residuals to fit
}}

\item{n}{\itemize{
\item number of rows to drop
}}

\item{debug}{\itemize{
\item flag to print diagnostics
}}
}
\value{
the trimmed dataframe
}
\description{
Trims rows from a dataframe based on a corresponding vector of "residuals".
}
