% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstackDataframe.R
\name{unstackDataframe}
\alias{unstackDataframe}
\title{Unstack columns in a dataframe by a set of factor levels}
\usage{
unstackDataframe(
  dfr,
  unstackCols = NULL,
  factorCols = NULL,
  dropCols = NULL,
  debug = TRUE
)
}
\arguments{
\item{dfr}{\itemize{
\item the dataframe
}}

\item{unstackCols}{\itemize{
\item the names of the columns to unstack
}}

\item{factorCols}{\itemize{
\item a vector of names of columns to use as factors
}}

\item{dropCols}{\itemize{
\item a vector of names of columns to drop from the result
}}

\item{debug}{\itemize{
\item flag (T/F) to print out debugging info
}}
}
\value{
dataframe with unstacked columns and factor levels
}
\description{
Creates a new dataframe by "unstack"ing columns in the
input dataframe by factor level combinations in a set of other columns. Uses
\code{getFactorLevels} to find the unique factor level combinations from
the set of dataaframe columns identified as factors.
}
\details{
None.
}
