% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_indices.R
\name{valid_indices}
\alias{valid_indices}
\title{Determine valid indices}
\usage{
valid_indices(idx, n)
}
\arguments{
\item{idx}{- vector of "candidate" indices}

\item{n}{- max valid index}
}
\value{
vector of valid indices, in order they occur in \code{idx}
}
\description{
Function to determine valid indices in a candidate vector
}
\details{
Valid indices into another vector are assumed to run from 1 to \code{n}. The
function returns the vector of valid indices, given \code{n}, in the input vector \code{idx}.
}
