% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTableToPDF.R
\name{writeTableToPDF}
\alias{writeTableToPDF}
\title{Create a pdf file from latex from a tabular object.}
\usage{
writeTableToPDF(
  tbl_tex,
  template_tex = NULL,
  pdf_name = "table.pdf",
  insertion_point = "&&tex_str",
  pageheight = 8.5,
  pagewidth = 6.5,
  margin = 0.1,
  orientation = "portrait",
  units = "in",
  cleanup = TRUE
)
}
\arguments{
\item{tbl_tex}{\itemize{
\item the tex string from tables::toLatex(tabular)
}}

\item{template_tex}{\itemize{
\item character string of tex template
}}

\item{pdf_name}{\itemize{
\item the desired name for the resulting pdf file
}}

\item{insertion_point}{\itemize{
\item string identifying point at which to insert the tex string into the template
}}

\item{pageheight}{\itemize{
\item page height (number, in units given by \code{units})
}}

\item{pagewidth}{\itemize{
\item page width (number, in units given by \code{units})
}}

\item{margin}{\itemize{
\item margin (number, in units given by \code{units})
}}

\item{orientation}{\itemize{
\item "portrait" or "landscape"
}}

\item{units}{\itemize{
\item units for measurements (e.g., "in" or "cm")
}}

\item{cleanup}{\itemize{
\item flag to clean up intermediate files
}}
}
\value{
Invisibly, the tex string for the pdf document.
}
\description{
Function to create a pdf file from latex from a tabular object.
}
\details{
This function inserts the latex string \code{tbl_tex} returned by \code{\link[tables:latex.tabular]{tables::toLatex()}} called on
a table \code{tbl}into a tex template (\code{tex_template}) for a document. It then converts the resulting tex to a pdf using
\code{\link[=writeTexToPDF]{writeTexToPDF()}}.

The insertion point in the template is indicated by the value of \code{insertion_point}. A default template is provided in
\code{tex/templateForTable.tex}.

The intermediate files are named "tex_DUMMY_TEX_FILE" with extensions .tex, .aux, .log, and .pdf. The pdf file
is copied to the file given by pdf_name (which defaults to 'table.pdf').
}
\note{
The use of orientation="landscape" reverses the sense of pageheight and pagewidth in determining the page size,
but does not rotate the table.
}
