% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTexToPDF.R
\name{writeTexToPDF}
\alias{writeTexToPDF}
\title{Create a pdf file from a tex string and tex template.}
\usage{
writeTexToPDF(
  template_tex_str = NULL,
  template_tex_file = NULL,
  tex_str = NULL,
  tex_file = NULL,
  insertion_point = "&&tex_str",
  pdf_name = NULL,
  cleanup = TRUE
)
}
\arguments{
\item{template_tex_str}{\itemize{
\item character string of tex template
}}

\item{template_tex_file}{\itemize{
\item name of text file with tex template
}}

\item{tex_str}{\itemize{
\item the tex string to insert into the template
}}

\item{tex_file}{\itemize{
\item the name of a tex file with the tex string
}}

\item{insertion_point}{\itemize{
\item string identifying point at which to insert the tex string into the template
}}

\item{pdf_name}{\itemize{
\item the desired name for the resulting pdf file
}}

\item{cleanup}{\itemize{
\item flag to clean up intermediate files
}}
}
\value{
Invisibly, the tex string used to create the pdf document.
}
\description{
Function to create a pdf file from tex.
}
\details{
This function inserts a tex string (in tex_str or tex_file) into a string version of a tex template
(template_tex_str) or file (template_tex_file) and then converts the resulting tex file to a pdf. The insertion
point in the template must be indicated by the value of \code{insertion_point}.

To use a 'template' without a '&&tex_str' insertion marker, leave tex_str and tex_file as NULLs.

The intermediate files are named "tex_DUMMY_TEX_FILE" with extensions .tex, .aux, .log, and .pdf. The pdf file
is copied to the file given by pdf_name (which defaults to the tex_file name with a pdf extension, or 'tex_str.pdf' if
tex_file is NULL).
}
