% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTransparency.R
\name{addTransparency}
\alias{addTransparency}
\title{Add transparency to a vector of colors}
\usage{
addTransparency(clrs, alpha = 1)
}
\arguments{
\item{clrs}{\itemize{
\item vector of colors
}}

\item{alpha}{\itemize{
\item level of transparency to add (0-1)
}}
}
\value{
corresponding vector of colors with transparency added
}
\description{
Function to add transparency to a vector of colors
}
\details{
None.
}
