% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateDataframe.R
\name{aggregateDataframe}
\alias{aggregateDataframe}
\title{Aggregate a dataframe according to "factor" levels.}
\usage{
aggregateDataframe(dfr, factors = NULL, vars = NULL)
}
\arguments{
\item{dfr}{- dataframe}

\item{factors}{- factors to aggregate by}

\item{vars}{- variable(s) to aggregate}
}
\value{
dataframe with vars aggregated by levels of "factor" columns
}
\description{
Function to aggregate a dataframe according to "factor" levels.
}
\details{
Uses sqldf.
}

