% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyCutPts.R
\name{applyCutPts}
\alias{applyCutPts}
\title{Determine bins by applying a vector of cutpoints}
\usage{
applyCutPts(zs, cutpts = seq(from = 25, to = 185, by = 5),
  truncate.low = TRUE, truncate.high = FALSE)
}
\arguments{
\item{zs}{- the vector to apply cutpoints to}

\item{cutpts}{- vector of cutpoints}

\item{truncate.low}{- vector of cutpoints}
}
\value{
a vector realigned to the cutpoints
}
\description{
Function to determine bins by applying a vector of cutpoints.
}
\details{
None.
}
