% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStats.dataframe.R
\name{computeStats.dataframe}
\alias{computeStats.dataframe}
\title{computeStats.dataframe: Calculate count, mean, stdev, and cv of selected columns of a dataframe}
\usage{
computeStats.dataframe(x, cols.factors = 1, cols.vars = 2)
}
\arguments{
\item{x}{\itemize{
\item a dataframe
}}

\item{cols.factors}{\itemize{
\item vector identifying columns to use as factors
}}

\item{cols.vars}{\itemize{
\item vector identifying columns to use as variables
}}
}
\value{
\itemize{
\item NULL (not yet implemented)
}
}
\description{
Returns count, mean, stdev, and cv of a dataframe
}
