% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeStats.vector.R
\name{computeStats.vector}
\alias{computeStats.vector}
\title{Returns count, mean, stdev, and cv of vector x}
\usage{
computeStats.vector(x)
}
\arguments{
\item{x}{- a vector of numbers (or object coerceable to a vector)}
}
\value{
- vector with count, mean, stdev and cv as named elements
}
\description{
Returns count, mean, stdev, and cv of vector x
}

