% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateCSVs.R
\name{concatenateCSVs}
\alias{concatenateCSVs}
\title{Concatenate csv files into one.}
\usage{
concatenateCSVs(caption = "Select csv files", out.csv = "test.csv",
  out.dir = NULL, method = "writeLines")
}
\arguments{
\item{caption}{- caption for file dialog box}

\item{out.csv}{- filename of output csv file}

\item{out.dir}{- directory for output csv file}

\item{method}{- specifies method used for concatenation ('writeLines' or 'write.table')}
}
\description{
Function to concatenate a number of csv files (with identical column strucuture)
into a single csv file.
}
\details{
None.
}
