% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCols.R
\name{deleteCols}
\alias{deleteCols}
\title{Delete columns in a dataframe or matrix}
\usage{
deleteCols(tbl, cols)
}
\arguments{
\item{tbl}{- dataframe or matrix from which to delete columns}

\item{cols}{- names of column to delete}
}
\value{
dataframe or matrix with columns removed
}
\description{
Function to delete columns in a dataframe or matrix
}
