% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMenu.R
\name{doMenu}
\alias{doMenu}
\title{Create a menu of items}
\usage{
doMenu(vals = NULL, prompts = NULL)
}
\arguments{
\item{vals}{\itemize{
\item vector of values
}}

\item{prompts}{\itemize{
\item ??
}}
}
\description{
Creates a menu of items from a vector of values and a vector of prompts
}
\details{
Uses package \code{uitls}.
}
