% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFile-FolderStructure.R
\name{getFileStructure}
\alias{getFileStructure}
\title{Get file structure}
\usage{
getFileStructure(
  path = ".",
  all.files = TRUE,
  full.names = TRUE,
  recursive = TRUE,
  include.dirs = FALSE
)
}
\arguments{
\item{path}{\itemize{
\item path to toplevel folder
}}

\item{all.files}{\itemize{
\item flag to include all files
}}

\item{full.names}{\itemize{
\item flag to return fullnames
}}

\item{recursive}{\itemize{
\item flag to determine structure recursively
}}

\item{include.dirs}{\itemize{
\item flag to include folders
}}
}
\value{
text vector
}
\description{
Get file structure
}
\details{
Uses \code{[base]{list.files}} to recursively determine file structure.
}
