% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFile-FolderStructure.R
\name{getFolderStructure}
\alias{getFolderStructure}
\title{Get folder structure}
\usage{
getFolderStructure(path = ".", full.names = TRUE, recursive = TRUE)
}
\arguments{
\item{path}{\itemize{
\item path to toplevel folder
}}

\item{full.names}{\itemize{
\item flag to return fullnames
}}

\item{recursive}{\itemize{
\item flag to determine structure recursively
}}
}
\value{
text vector
}
\description{
Get folder structure
}
\details{
Uses \code{[base]{list.dirs}} to determine folder structure.
}
