% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFunctionName}
\alias{getFunctionName}
\title{Get the name of a function}
\usage{
getFunctionName(fcn)
}
\arguments{
\item{fcn}{- function name as an expression}
}
\value{
function name as character string
}
\description{
Returns the name of the input function as a character string
}

