% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObj_saveObj.R
\name{getObj}
\alias{getObj}
\title{Get saved object from an RData file}
\usage{
getObj(fn)
}
\arguments{
\item{fn}{- filename from which to extract R object}
}
\value{
the object
}
\description{
Function to get a saved object ('obj') from an RData file.
}
\details{
The original R object must have been saved with the name 'obj' (see
\link{\code{wtsUtilities::saveObj}} for a function to do so).
}
