% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOperatingSystem.R
\name{getOperatingSystem}
\alias{getOperatingSystem}
\title{Get the operating system name.}
\usage{
getOperatingSystem()
}
\value{
Either 'MacOSX' or 'Windows'
}
\description{
Function that returns the operating system name.
}
\details{
None.
}
