% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_FixOnePlotDim.R
\name{gg_FixOnePlotDim}
\alias{gg_FixOnePlotDim}
\title{Calculate the missing dimension of a ggplot2 object given the other}
\usage{
gg_FixOnePlotDim(
  ggobj,
  fxdDim,
  fitWidth = TRUE,
  units = "in",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{ggobj}{\itemize{
\item a ggplot2 object (or possibly a gtable)
}}

\item{fxdDim}{\itemize{
\item value for the fixed dimension
}}

\item{fitWidth}{\itemize{
\item flag indicating fixed dimension is width
}}

\item{units}{\itemize{
\item character string with units for the fixed ddimension
}}

\item{verbose}{\itemize{
\item flag to print intermediate information
}}

\item{...}{\itemize{
\item not used
}}
}
\value{
a 2-element list (height, width) with the plot dimensions
}
\description{
Function to calculate the missing dimension of a
ggplot2 object given the other dimension.
}
\details{
Based on code from the DeLuciatoR::get_dims function.

This function
calculates the total height/width based on setting the total width/height (the
latter determines the size of a "null" unit for expandable grobs).
Another approach is taken in \link{gg_GetPlotDims}, which fixes the size of null units,
then calculates the resulting height and width.
}
