% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_SavePlotWithOneFixedDim.R
\name{gg_SavePlotWithOneFixedDim}
\alias{gg_SavePlotWithOneFixedDim}
\title{Save a ggplot2 plot to a file, specifying the size of one dimension}
\usage{
gg_SavePlotWithOneFixedDim(
  filename,
  plot = last_plot(),
  fxdDim,
  fitWidth = TRUE,
  units = "in",
  device = NULL,
  dpi = 200
)
}
\arguments{
\item{filename}{\itemize{
\item file name for output
}}

\item{plot}{\itemize{
\item ggplot2 plot object to save (defaults to last plot)
}}

\item{fxdDim}{\itemize{
\item value for the fixed dimension
}}

\item{fitWidth}{\itemize{
\item flag indicating fixed dimension is width (default is TRUE)
}}

\item{units}{\itemize{
\item character string with units for the fixed dimension
}}

\item{device}{\itemize{
\item graphics device to print to
}}

\item{dpi}{\itemize{
\item resolution (default=200 dpi)
}}
}
\description{
Function to save a ggplot2 plot to a file,
specifying the size of one dimension.
}
\details{
Uses \link{gg_FixOnePlotDim} to determine the value of the missing plot dimension.
}
