% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/keepCols.R
\name{keepCols}
\alias{keepCols}
\title{keepCols: keep columns from a dataframe}
\usage{
keepCols(tbl, cols, debug = FALSE)
}
\arguments{
\item{tbl}{- dataframe with columns to keep or discard}

\item{cols}{- vector of column names to keep}

\item{debug}{- flag (T/F) to print debugging info}
}
\value{
a dataframe with columns to be kept, or NULL if no column names matched.
}
\description{
Create dataframe with subset of columns of input dataframe "tbl" by
      keeping only columns matching the specified names.
}
\details{

}

