% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDays.R
\name{parseDays}
\alias{parseDays}
\title{Convert character vector of dates to numeric vector of days-of-month.}
\usage{
parseDays(x, format = "MM/DD/YYYY")
}
\arguments{
\item{x}{- the character vector to convert}

\item{format}{- the date format (e.g. 'MM/DD/YYYY','DD-MON-YYYY')}
}
\value{
numeric vector of day of month (1-31)
}
\description{
Function to convert a character vector of dates to a numeric vector of days-of-month.
}

