% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMonths.R
\name{parseMonths}
\alias{parseMonths}
\title{Convert character vector of dates to numeric vector of months.}
\usage{
parseMonths(x, format = "MM/DD/YYYY")
}
\arguments{
\item{x}{- the character vector to convert}

\item{format}{- the date format (e.g. 'MM/DD/YYYY','DD-MON-YYYY')}
}
\value{
numeric vector of months (1-12)
}
\description{
Function to convert a character vector of dates to a numeric vector of months.
}

