% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TextFunctions.R
\name{parseNum}
\alias{parseNum}
\title{Parse a numeric expression from a character string.}
\usage{
parseNum(str)
}
\arguments{
\item{str}{- character string to parse}
}
\value{
parsed and evaluated numerical expression
}
\description{

}

