% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextFunctions.R
\name{parseNum}
\alias{parseNum}
\title{Parse numeric expressions from a character vector.}
\usage{
parseNum(str)
}
\arguments{
\item{str}{- character vector to parse}
}
\value{
parsed and evaluated numerical expressions as a vector
}
\description{
Function to parse numeric expressions from a character vector.
}
\details{
None.
}
