\name{parseYears}
\alias{parseYears}
\title{Convert character vector of dates to numeric vector of days-of-month.}
\usage{
parseYears(x, format = "MM/DD/YYYY", sep = "/")
}
\arguments{
  \item{x}{- the character vector to convert}

  \item{format}{- the date format (e.g. 'MM/DD/YYYY')}

  \item{sep}{- the date-month-yar divider (e.g. '/')}
}
\value{
numeric vector of months (1-12)
}
\description{

}

