% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rootMedianSquare.R
\name{rootMedianSquare}
\alias{rootMedianSquare}
\title{Compute the root median square deviation (rmsd) of a vector}
\usage{
rootMedianSquare(y, na.rm = TRUE)
}
\arguments{
\item{y}{- vector to compute rmsd for}

\item{na.rm}{- flag to remove NAs before calculating}
}
\value{
the rmsd, or NA if y is all NAs
}
\description{
Function to compute the root median square deviation (rmsd) of a vector
}
\details{

}

