% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectFile}
\alias{selectFile}
\title{Select a file.}
\usage{
selectFile(ext = "*", caption = paste("Select .", ext, " file to import",
  sep = ""))
}
\arguments{
\item{ext}{- extension for files to choose from}

\item{caption}{- caption for file dialog (if file name not provided)}
}
\value{
Selected file name. Returns NULL if the user canceled selection using the file dialog.
}
\description{
Function allows the user to select a file using a gui interface.
}

