% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFile.R
\name{selectFile}
\alias{selectFile}
\title{Select a file}
\usage{
selectFile(
  ext = "*",
  caption = paste("Select .", ext, " file(s) to import", sep = ""),
  multi = FALSE
)
}
\arguments{
\item{ext}{\itemize{
\item extension for files to choose from
}}

\item{caption}{\itemize{
\item caption for file dialog (if file name not provided)
}}

\item{multi}{\itemize{
\item flag (T/F) to allow multiple file selection
}}
}
\value{
Selected file name(s). Returns NULL if the user canceled selection using the file dialog.
}
\description{
Function allows the user to select a file using a gui interface.
}
\details{
Uses \code{tcltk::tk_choose.files} to select file(s).
}
