% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector by converting to z-scores.}
\usage{
standardize(x)
}
\arguments{
\item{x}{- vector to standardize}
}
\value{
standardized vector as z-scores (same size as x)
}
\description{
Function to standardize a vector by converting to z-scores.
}
\details{

}

