% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector or columns in a dataframe by converting to z-scores.}
\usage{
standardize(x, cols = NULL)
}
\arguments{
\item{x}{- vector or dataframe to standardize}

\item{cols}{- vector of column names to standardize (if x is a dataframe)}
}
\value{
standardized vector as z-scores (same size as x), or dataframe
}
\description{
Function to standardize a vector or columns in a dataframe by converting to z-scores.
}
