% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.vector}
\alias{standardize.vector}
\title{Standardize a vector by converting to z-scores.}
\usage{
standardize.vector(x)
}
\arguments{
\item{x}{\itemize{
\item vector to standardize
}}
}
\value{
standardized vector as z-scores (same size as x)
}
\description{
Function to standardize a vectorby converting to z-scores.
}
\details{
If sd(x) is 0, then 0*x is returned. If sd(x) is NA, then NA+x is returned.

not exported
}
