% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum.R
\name{Sum}
\alias{Sum}
\title{Sum with NAs removed}
\usage{
Sum(x, verbose = FALSE)
}
\arguments{
\item{x}{- object to be summed}

\item{verbose}{- flg to print diagnostic info}
}
\description{
Convenience function to sum with NAs removed.
}
\details{
If x is numeric, then \code{Sum(x)} is equivalent to \code{sum(x,na.rm=TRUE)}. If x is a character vector,
then \code{Sum(x)} returns x[1]. If x is logical, then \code{Sum(x)} is equivalent to \code{any(x,na.rm=TRUE)}
}
