% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTableToPDF.R
\name{writeTableToPDF}
\alias{writeTableToPDF}
\title{Create a pdf file from latex from a tabular object.}
\usage{
writeTableToPDF(
  tbl_tex,
  template_tex = NULL,
  pdf_name = "table.pdf",
  insertion_point = "&&tex_str",
  pageheight = 8.5,
  pagewidth = 6.5,
  margin = 0.1,
  orientation = "portrait",
  units = "in",
  cleanup = FALSE
)
}
\arguments{
\item{tbl_tex}{- the tex string from tables::toLatex(tabular)}

\item{pdf_name}{- the desired name for the resulting pdf file}

\item{insertion_point}{- string identifying point at which to insert the tex string into the template}

\item{pageheight}{- page height (number, in units given by \code{units})}

\item{pagewidth}{- page width (number, in units given by \code{units})}

\item{margin}{- margin (number, in units given by \code{units})}

\item{orientation}{- "portrait" or "landscape"}

\item{units}{- units for measurements (e.g., "in" or "cm")}

\item{cleanup}{- flag to clean up intermediate files}

\item{template_tex_str}{- character string of tex template}
}
\value{
Invisibly, the tex string for the pdf document.
}
\description{
Function to create a pdf file from latex from a tabular object.
}
\details{
This function inserts the latex string \code{tbl_tex} returned by [tables::toLatex(tbl)] called on
a table \code{tbl}into a tex template (\code{tex_template_str}) for a document. It then converts the resulting tex to a pdf using
[writeTexToPDF()].

The insertion point in the template is indicated by the value of \code{insertion_point}. A default template is provided in
\code{tex/templateForTable.tex}.

The intermediate files are named "tex_DUMMY_TEX_FILE" with extensions .tex, .aux, .log, and .pdf. The pdf file
is copied to the file given by pdf_name (which defaults to 'table.pdf').
}
\note{
The use of orientation="landscape" reverses the sense of pageheight and pagewidth in determining the page sie,
but does not rotate the table.
}
